发布

+++ install.sh
@@ -18,21 +18,10 @@ readonly repodir="/yzaj/${REPO}"
 readonly tempdir="/yzaj/temp/${REPO}"
 readonly src="/etc/profile.d/yzaj-${REPO}.sh"
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+if [[ ! -f "/${MINTTY}" ]]; then
+  err "Please use cygwin to execute"
+  exit "${E_USE_MINTTY}"
+fi
 
 if ! type "${CMD}"; then
   apt-cyg install "${CMD}"
@@ -41,8 +30,8 @@ fi
 git clone --depth 1 "${url}" "${repodir}"
 
 if [[ -d "${repodir}/bin" ]]; then
-  chmod -R 755 "${repodir}"/bin/
+  chmod -R 755 "${repodir}"/bin
   echo "export PATH=\${PATH}:${repodir}/bin" > "${src}"
 fi
 
-mkdir -p "/${OWNER}"/log/"${REPO}"
+mkdir -p "${tempdir}"
